+++ src/components/status/mod.rs
@@ -1,7 +1,9 @@
+mod pressed_key;
 mod span;
 mod spinner;
 mod status_phantom;
 
-pub(crate) use spinner::StatusSpinner;
+pub(crate) use pressed_key::PressedKey;
 pub(crate) use span::StatusSpan;
+pub(crate) use spinner::StatusSpinner;
 pub(crate) use status_phantom::DbResponseStatusListener;